TypeError: val is null
TypeError: val is null in for ( var j=0, jLen=val.length ; j<jLen ; j++ )
The function `-api _fnSetObjectDataFn( mSource )` returns a function that can be used to set data from a source object, nevertheless it doesn't handle correctly the case when the value is null `(val === null)`, which is a legitimate value. This can cause problems for function relaying on this function. One example is when Editor use this function to set the values `-api Editor.prototype._submit`, if a user didn't select anything in a multiple value select, the value for this field is null. A very simple way (but probably not the most elegant) to solve this one is by ignoring a null value passed and early returning.