Let me explain our application and why we need this feature.
We search data from network devices and the data from these devices is stored in elasticsearch. A search (filter) query can be sent to find information (terms/keywords), but elasticsearcg also has the ability to say “not these terms” in a query as well. As such, our UI would be be designed with a query filter input and a "NOT these terms" filter input. So if you will, two global filter inputs. One input for “find this stuff” and the other input for “but not this stuff” or exclude this stuff in the results.

I have made changes to the code so that this information comes through to the server data function like so…

DEVICEREPORTS.setElasticFnServerData = function (type, indice_pattern, deviceCustomString2) {

    var recordType = type,
        service = deviceCustomString2;

    this.setFnServerData(function (sSource, aoData, fnCallback, oSettings) {
        var body,
            builder,
            draw = oSettings.iDraw,
            length = oSettings._iDisplayLength,
            start = oSettings._iDisplayStart,
            globalSearch = oSettings.oAjaxData.search.value,
            globalExclude = oSettings.oAjaxData.exclude.value,
            columnSearches = aoData[1].value,

These changes add the functionality needed for a server data function to work.
I need to learn more about the code base to make this feature work client side too.