Build: Changing how builds are created to allow ES module files to be created as well as the standard JS files.
No effect on the JS files that are generated. They are functionally identical to what they were before. However, the source files are simplified as they no longer need to contain a UMD wrapper - that is added at build time (or ESM wrapper). Note that the build process for DataTables core is slightly different than what is used for the extensions.

Also switching to use uglify-js as the minifier as it works well with ES modules, while Closure Compiler attempts to resolve and inline them.