Update: Allow CommonJS loader to operate without needing to execute a factory function when a window object is available
The goal here is to allow easier use with `import DataTable from
'datatables.net';` when that statement is down compiled to a
`require()`. The problem with that is that it is a CommonJS loader,
which has a different syntax from the ESM, so if you are using ESM you
do:
  import DataTable from 'datatables.net';

But if you are using CommonJS:
  import DTfactory from 'datatables.net';
  const DataTable = DTfactory();

The real killer is that many don't know if they are using CommonJS or
ESM. If the down compile to CommonJS wasn't a thing, this wouldn't be a
problem, but since they are, those two first lines are identical and if
you don't know which loader you are using, then all bets are off for
what is loaded.

The aim here is to allow the CommonJS loader to operate without need to
be executed. That is complicated by jQuery which returns a factory
function, but only when a window object is not present. We check for
that, and retain the old behaviour if that is the case.

If the window is present, then we want to return the DataTable object.
That now happens and the DataTables function will check to see if it is
being used as a factory. If it is, then it just returns the DataTable
object, having set the window and $ variables as needed.