Dev: Adding cache for rendered data
Couple of reasons for this:

1. 2 changed the column width calculations to always run, resulting in
   the rendering function running twice for `display` on start up (the
   actual display and the width string). This triggered an error in the
   unit tests since the rendering function was called more than is
   needed. Indeed, I think it is correct that the `display` only be
   called when actually needed for display.

2. Performance. Less calls to the rendering function (at the expense of
   another cache array in RAM).

I've also removed the `td` element part of the column width calculations
and cloning. That appears to go all the way back to 2013 and before. I
suspect it is a left over from when I actually tried moving the elements
around for the width. No need to access the contents of the cell and
then dump it into a new cell. Just use the cached value now.