Dev fix: I'd broken the assignment of `columnDefs` using a class name due to the header detection and the removal of `.nTh`.
Addressed by passing in the header as it is detected when the table init
starts up. If there isn't a header element at this stage, then the
header layout just resolves to be an empty array and no column could be
targeted via a class since there is no class assignment yet!