Dev - update: `-init layout` change to simpiler objects
Attempting to simplify how features can be specified via objects. Use
the property names as the feature and its value to be passed to the
feature. Normally this will be a list of options, but it could be a
simple string or any other value.

This will also make the typescript integration easier rather than
attempting to mix all possible feature options together.