New: `-api column().header()` and `-api column().footer()` (and their plurals) can specify a row index in the header / footer which the API method should return. This provides support for multi-row headers and footers via the API for the first time in DataTables.
Note that the legacy `orderCellsTop` parameter will cause the top cell
to be selected if that parameter is `true` and no parameter is given to
the method. If not specified, the default behaviour is to select the
bottom row in the header, the top row in the footer.

Dev: Reading of multiple header rows was overwriting the title with what
had been read before, since it thought a user value had been given. The
`columns.title` option will only set the title on the first row.
Anything else needs to be done using a provided header.