New: `data-dt-order` (attribute) can be applied to any cell or row in the header of the table to control if the order click listener is attached to the cell or not. It can also be used to control the ordering status icons.
Deprecated: `-init orderCellsTop` is now superseeded by `data-dt-order`
which offers much more granular control. It still works, but is slightly
different from v1 where it was `false` by default. Now it is `null` by
default so the listener and icons are added to add cells in the header
(unless `data-dt-order`) is used. Support for this property is still
available though - it can be set to `true` or `false` to match its
behaviour in 1.x.