New: `-api column().title()` and `-api columns().titles()` to get / set column titles, including support for multiple row headers. Note that this requires that DataTables wraps the text content of the header with a `span` element. It has a class of `dt-column-title`.
I did look into the option of having the title operate on the first text
node in the header cell, but that would then preclude the title ever
having HTML content, which I think it is okay to do (spans for
sub-titles for example). A `span` (or some other element, but span being
neutral and semantically meaningless is perfect) is used to be able to
modify the title on the fly.

Tests added. A few needed to be updated to account for the change in the
HTML structure.

DD-566