Fix: Plural methods were available on singular API invokations. For example `row().nodes()` was available, even although there was no definitation for it - it was `rows().nodes()` leaking through.
Couple of tests were falling into this trap and needed to be updated.

I considered changing `eq()` or adding a `clone()` method to the API to
basically expose the functionality of `_selector_first()`, which
basically creates a new API instance and assigned a value to it, but but
it might have been confusing if you'd cloned a `rows()` method - which
wouldn't have the rows()` child methods on it - it would always be the
base.