Performance: Significant performant improvement in `-api rows().every()`, `-api columns().every()` and `-api cells().every()`.
The API instance for the singular form is now reused for each item,
rather than creating a new instance for every item. This means that only two
API instances are created in the minimal case, rather than items+1.

As an example - consider:

```
for (var i=0 ; i<10000 ; i++) {
  new DataTable.Api('#example').rows().every(function () {
    this.data();
  });
}
```

On Firefox, before this optimisation it took a little over ten seconds
to run. With this change it is 600mS.