New: `-init columns.footer` option which can be used to create a column footer cell if there isn't one present in the DOM, or set the text of an existing cell.
Updated: CSS `no-footer` class has been removed from the table for when
the table does not have a `-tag tfoot` element. DataTables will now
insert an empty `-tag tfoot` if there isn't on in the table, which
significantly simplies the logic for scrolling and captions. It also
simplies the CSS for table borders, and the CSS has been updated to
reflect this.

DD-551