New - examples: Ability to run the examples with any styling framework.  This is done with a menu at the top left of the examples letting you select the styling framework to use for the example (matching how the initialisation code style and colour scheme can already be chosen). Some examples can still force a specific styling framework (e.g. the Bootstrap5 specific style).
Note that Firefox has a FOUC with all of these examples which I've
reported ( https://bugzilla.mozilla.org/show_bug.cgi?id=1851842 ).
Chrome does not exhibit this issue (it is almost a show stopper, but I
think that this ability is quite useful for debugging as well as to
demonstrate the styling flexibility).

DD-2413