New: Diacritic search support. You can now use DataTables search to look for accented characters without typing the accent. For example searching for `creme brulee` will match `Crème Brulée` in the table. New: `-api DataTable.util.diacritics()` method exposing the diacritic stripping function we use, and providing the ability to replace it if you require.
ALLAN - RELEASE NOTES!
This also required reverting the following:
Revert "Update: Smart search new matches at from the start of a word rather than in the middle. E.g. "ingeer" will not match "Engineer". This is to stop false positives such as a search for "male" also matching "female"."

This reverts commit ff8b4e3fb356aa3646dcf8f26f769d4875a9a889.

This is because `\b` doesn't allow for diacritics at the start of a
word. I've also realised that it wouldn't allow search to happen
correctly in Chinese and other languges where whitespace isn't
important.

DD-481
DD-223