Fix: Update Material example to use 14.0.0
The input element and select aren't right in the Metrial world, but they
each require a huge amount of extra markup which isn't possible with the
DataTables defaults. It would be possible to use the feature plug-ins to
replace the defauls, but until someone actually asks for this, I don't
see any benefit in doing it. I find the Material documentation
exceptionally difficult to use and it isn't a particularly enjoyable
developer experience.