Dev: More tweaks to attempt to support cash
There are a bunch of small things that cash doesn't support that make
supporting it non-trivial. Most we can overcome by simply trying it and
then tweaking - for example not supporting an array of nodes for
`.append()` or using attributes in the second parameter when creating
elements.

However, there are two showstoppers:

1. No `$.Event` constructor or `triggerHandler` method. We depend upon
   these in our events and they would need to be added.

2. No `$.ajax`, which would mean everything needs to be done in a
   function with `fetch` or we implement one.

I really want to support cash because 1, it is tiny! and 2, I'm going to
get a shit lot of comments about how out of date jQuery is when DT2
lands. However, at the moment cash doesn't quite fill the jquery gap.