Reinstate: The `dom` option is now available again and will be used if set. DON'T USE IT! It will be removed in DT3.
I've managed to get the addition of this feature down to 415 additional
bytes (minified) which I think is worth it to provide the upgrade path.
The `dom` parameter wasn't depricated in any 1.x release as there was no
replacement then. This provides the 2.x series as a change over from
`dom` to `layout`.

Tests reinstated and docs updated