Fix: Remove blur from click events, allowing focus to remain on the selected element. Browser's no longer show a focus ring on click which is why this was originally there.
Many thanks for Jacques Newman and Kelly Yokote from the MS Edge team
for working with me on this one. The original issue relates to how
"Narrator" works in Windows and it conflicting with DataTables' own
efforts at accessability. Narrator will change keyboard events into
mouse events (presumably to allow for applications which haven't taken
keyboard navigation into account), but DataTables event function assumed
that if a mouse event happened, then it was a mouse triggering it, and
thus it would be safe to blur. That isn't the case with Narrator on.

The reason to blur was back around the IE6 days browsers would show a
focus ring for items that have focus. This isn't attractive to a mouse
user, so a blur or some other techinique to fill the focus outline was
typically used.

That is no longer the case. Jacques pointed out this part of MDN:
https://developer.mozilla.org/en-US/docs/Web/CSS/:focus-visible . On
click the focus is given to the element, but it is not shown visibly.
Thus there is no need for the blur any more, allowing keyboard and mouse
interaction to work together and Narrator to work as planned.