Fix: Don't modify the display array in `-api row().remove()` - it should wait for the next draw
This relates to https://github.com/DataTables/DataTablesSrc/issues/269
and is fairly nuanced - it boils down to the fact that the display array
should reflect what is shown in the table, and deleting a value from it
will cause an index offset. The other option might be to make the
element null, as happens with the master, but that would be a change in
type and I'm not certain everything could handle that. The array will
regenerate on the next draw, so it is properly in sync with what is
drawn now.