New: Ability to control what buttons appear in the paging control through three new options for the feature: * `-feature paging.numbers` * `-feature paging.previousNext` * `-feature paging.firstLast`
This provides similar functionality to `-feature paging.type` but is more intuitive. It also allows the paging feature code to be more compartmentalised, and will allow a reduction in code size when the deprecated properties are removed.

Deprecated: `-init pagingType` and `-feature paging.type` are now redundant and will be removed in DataTables 3. If you are using these properties it is recommended that you update them when possible.