Fix: The automatic resize handling wouldn't allow the content to get smaller.
The fix took a lot longer than what the title suggests, but it basically
boils down to the fact that when you call `.observe()` it will
immediately trigger a callback (if the element is visible!) which I very
much did not expect. Accounting for that allows things to work as
expected.