Dev: Continued TS improvements and remove jQuery dependency
Focus here on core/columns.ts.

Small update to the tests as `:gt()` is a jQuery selector, not CSS, and
so it would throw an error if used now.

BACKWARDS COMPATIBILITY NOTE: If using `-init columns.columnDefs` with a
jQuery specific selector (e.g. `:eq()`, `:gt()` or `:lt()` you will need
to update it to be a CSS selector.