Dev: Updating constructor for TS and jQuery-less
Needed a way to allow DataTables to be attached to a jQuery object after
initial load. To be honest, v2 needs this as well, but it does an
`import $ from 'jquery';` so it has never been an issue. That
requirement required a little restructing since `DataTable` needs to be
imported for the setup function.

I've also merged the constructor in to the core, rather than having it
in a separate function.