Dev: Continued WIP for Api using prototypes
Can't use Typescript classes as it will compile things like
`lastIndexOf = Array.prototype.lastIndexOf` to `this.lastIndexOf = ...`
- i.e. it is a property, not assigned to the prototype, which is what I
need here. So refactored `Api` to be a function again here.

The current code appears to be okay for the methods, but doesn't work
for property based methods such as `columns.adjust()`. TBD