Dev: Api back to working status, using prototypes for methods, and property extending similar to before for the properties.
This is because the scope for the nested properties couldn't otherwise
be set from the prototype. It is possible to use defineProperty, but
that would only work for cases where there isn't already a method. This
hybrid approach appears to be the way to do it.

For `rows().every()` there is about a 50% speed improvement
For `cells().every()` it is about 400% slower! To be looked at