Dev: Fix issue with multiple API instances and tables
This was a bit of a nasty one - I had it writing the property methods
(e.g. page.info()) to Api.prototype.page.info, which of course is shared
between all instances, so the scope would be for which every API
instance was last contstructed (extended).

The fix is to break that sharing. The instance needs its own object,
rather than using the prototype method.