Dev: Revising how the scoping for the API methods scope worked
Doing something like `table.column().draw()` is valid (if redundant) but
an error was being thrown because the scope isn't 1:1 compatiable. I
need the scope of the functions at the top of the chain to changes, and
this introduction of `ApiScopeable` does that (name might change).

`table.table().header.strcuture()` still not right - tbd.