Dev: Starting work on removing the internal use of hungarian notation
Since 1.10 camelCase has been the form used and prompted by the
documentation, but internally I just mapped the new style to the old.

Now the intention is the reverse, to have the core operate only with
camelCase, but input configuration options with Hungraian notation
parameters to be mapped to camelCase.